+++ CHANGES.rst
@@ -14,6 +14,7 @@ Improvements:
  * Key backup: Implement setup reminder (#2211).
  * Key backup: Implement recover reminder (#2206).
  * Key backup: Update key backup setup UI and UX (PR #2243).
+ * Key backup: Logout warning (#2245).
  
 Bug fix:
  * Use white scroll bar on dark themes (#2158).